requirements.txt
asgiref==3.5.2
Django==3.2.10
django-registration-redux==2.10
django-versatileimagefield==2.2
Pillow==9.1.1
psycopg2-binary==2.9.1
python-decouple==3.6
python-magic==0.4.27
pytz==2022.5
six==1.16.0
sqlparse==0.4.3
